Modifications du package Categorie Utilisation du Mappage et des DTO avec le package Evenement

+++ src/Web2.API/Controllers/CategoriesController.cs
@@ -17,28 +17,28 @@ namespace Web2.API.Controllers
     public class CategoriesController : ControllerBase
     {
         private readonly ICategoryBL _categoryBL;
-        private readonly TP2A_Context _context;
         private readonly IMapper _mapper;
 
-        public CategoriesController(ICategoryBL categoryBL, TP2A_Context context, IMapper mapper)
+        private readonly TP2A_Context _context;
+
+        public CategoriesController(ICategoryBL categoryBL, IMapper mapper, TP2A_Context context)
         {
             _categoryBL = categoryBL;
-            _context = context;
             _mapper = mapper;
+            _context = context;
         }
 
         // GET: api/<CategoriesController>
         [HttpGet]
         [ProducesResponseType(typeof(List<CategorieDTO>), StatusCodes.Status200OK)]
-        public async Task<IEnumerable<CategorieDTO>> Get(int pageIndex = 1, int pageCount = 5)
+        public async Task<IEnumerable<CategorieDTO>> Get()
         {
-            var categories = await _context.Categories.Include(std => std.Evenements)
-                .Skip((pageIndex - 1) * pageCount)
-                .Take(pageCount)
-                .AsNoTracking()
+            var categories = await _context.Categories
+                .Include(std => std.Evenements)
                 .ToListAsync();
 
-            return _mapper.Map<IList<Categorie>, IEnumerable<CategorieDTO>>(categories);
+            var categoriesDTO = _mapper.Map<IEnumerable<CategorieDTO>>(categories);
+            return categoriesDTO;
         }
 
         // GET api/<CategoriesController>/5
@@ -49,7 +49,8 @@ namespace Web2.API.Controllers
         {
             var categorie = await _context.Categories.Include(std => std.Evenements).FirstOrDefaultAsync(x => x.ID == id);
 
-            return _mapper.Map<CategorieDTO>(categorie);
+            var categorieDTO = _mapper.Map<CategorieDTO>(categorie);
+            return categorieDTO;
         }
 
         // POST api/<CategoriesController>
@@ -57,10 +58,14 @@ namespace Web2.API.Controllers
         [ProducesResponseType(typeof(CategorieDTO), StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [Consumes(MediaTypeNames.Application.Json)]
-        public async Task Post([FromBody] CategorieDTO categorie)
+        public async Task<IActionResult> Post([FromBody] CategorieDTO categorieDTO)
         {
-            _context.Categories.Add(new Categorie { ID = categorie.ID, Name = categorie.Name, });
+            var categorie = _mapper.Map<Categorie>(categorieDTO);
+            _context.Categories.Add(categorie);
             await _context.SaveChangesAsync();
+
+            var CategorieDTOCreee = _mapper.Map<CategorieDTO>(categorie);
+            return CreatedAtAction(nameof(Get), new { id = CategorieDTOCreee.ID }, CategorieDTOCreee);
         }
 
         // PUT api/<CategoriesController>/5
@@ -68,25 +73,40 @@ namespace Web2.API.Controllers
         [ProducesResponseType(typeof(CategorieDTO), StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [Consumes(MediaTypeNames.Application.Json)]
-        public async Task Put(int id, [FromBody] CategorieDTO categorie)
+        public async Task<IActionResult> Put(int id, [FromBody] CategorieDTO categorieDTO)
         {
-            var categorieAMAJ = _context.Categories.Include(std => std.Evenements).FirstOrDefault(x => x.ID == id);
+            var categorieAMAJ = await _context.Categories.Include(std => std.Evenements).FirstOrDefaultAsync(x => x.ID == id);
 
-            categorieAMAJ.ID = categorie.ID;
-            categorieAMAJ.Name = categorie.Name;
-
-            await _context.SaveChangesAsync();
+            if (categorieAMAJ != null)
+            {
+                _mapper.Map(categorieDTO, categorieAMAJ);
+                await _context.SaveChangesAsync();
+                return Ok(categorieDTO);
+            }
+            else
+            {
+                return NotFound();
+            }
         }
 
         // PUT api/<CategoriesController>/5
         [HttpDelete("{id}")]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
         [ProducesResponseType(StatusCodes.Status409Conflict)]
-        public async Task Delete(int id)
+        public async Task<IActionResult> Delete(int id)
         {
-            var categorie = _context.Categories.Find(id);
-            _context.Categories.Remove(categorie);
-            await _context.SaveChangesAsync();
+            var categorie = await _context.Categories.FindAsync(id);
+
+            if (categorie != null)
+            {
+                _context.Categories.Remove(categorie);
+                await _context.SaveChangesAsync();
+                return NoContent();
+            }
+            else
+            {
+                return Conflict();
+            }
         }
 
     }
